Moved product model to its own folder.
