+++ PhotoPlusFrontend/PhotoPlus/src/app/services/product/product.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http'
-import { Product } from './product';
+import { Product } from '../../models/product/product';
 import { Category } from '../../models/category/category';
-import { element } from 'protractor';
 
 @Injectable({
   providedIn: 'root'
